[hlm-621]: Refactored as per review comments
