+++ health-services/individual/src/main/java/org/egov/individual/Constants.java
@@ -3,4 +3,8 @@ package org.egov.individual;
 public interface Constants {
     String SET_INDIVIDUALS = "setIndividuals";
     String VALIDATION_ERROR = "VALIDATION_ERROR";
+    String SYSTEM_GENERATED = "SYSTEM_GENERATED";
+    String GET_INDIVIDUALS = "getIndividuals";
+    String GET_IDENTIFIER_TYPE = "getIdentifierType";
+    String GET_ID = "getId";
 }
