+++ health-services/individual/src/main/java/org/egov/individual/validators/IsDeletedSubEntityValidator.java
@@ -1,7 +1,7 @@
-package org.egov.individual.service;
+package org.egov.individual.validators;
 
 import org.egov.common.models.Error;
-import org.egov.common.utils.Validator;
+import org.egov.common.validator.Validator;
 import org.egov.individual.web.models.Address;
 import org.egov.individual.web.models.Identifier;
 import org.egov.individual.web.models.Individual;
