+++ health-services/individual/src/main/java/org/egov/individual/validators/NonExistentEntityValidator.java
@@ -1,8 +1,8 @@
-package org.egov.individual.service;
+package org.egov.individual.validators;
 
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.models.Error;
-import org.egov.common.utils.Validator;
+import org.egov.common.validator.Validator;
 import org.egov.individual.repository.IndividualRepository;
 import org.egov.individual.web.models.Individual;
 import org.egov.individual.web.models.IndividualBulkRequest;
@@ -17,23 +17,25 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static org.egov.common.utils.CommonUtils.getEntitiesWithMismatchedRowVersion;
+import static org.egov.common.utils.CommonUtils.checkNonExistentEntities;
 import static org.egov.common.utils.CommonUtils.getIdFieldName;
-import static org.egov.common.utils.CommonUtils.getIdMethod;
 import static org.egov.common.utils.CommonUtils.getIdToObjMap;
+import static org.egov.common.utils.CommonUtils.getMethod;
+import static org.egov.common.utils.CommonUtils.getObjClass;
 import static org.egov.common.utils.CommonUtils.notHavingErrors;
 import static org.egov.common.utils.CommonUtils.populateErrorDetails;
-import static org.egov.common.utils.ValidatorUtils.getErrorForRowVersionMismatch;
+import static org.egov.common.utils.ValidatorUtils.getErrorForNonExistentEntity;
+import static org.egov.individual.Constants.GET_ID;
 
 @Component
-@Order(value = 5)
+@Order(value = 4)
 @Slf4j
-public class RowVersionValidator implements Validator<IndividualBulkRequest, Individual> {
+public class NonExistentEntityValidator implements Validator<IndividualBulkRequest, Individual> {
 
     private final IndividualRepository individualRepository;
 
     @Autowired
-    public RowVersionValidator(IndividualRepository individualRepository) {
+    public NonExistentEntityValidator(IndividualRepository individualRepository) {
         this.individualRepository = individualRepository;
     }
 
@@ -41,18 +43,19 @@ public class RowVersionValidator implements Validator<IndividualBulkRequest, Ind
     @Override
     public Map<Individual, List<Error>> validate(IndividualBulkRequest request) {
         Map<Individual, List<Error>> errorDetailsMap = new HashMap<>();
-        Method idMethod = getIdMethod(request.getIndividuals());
-        Map<String, Individual> iMap = getIdToObjMap(request.getIndividuals().stream()
-                .filter(notHavingErrors())
-                .collect(Collectors.toList()), idMethod);
+        List<Individual> individuals = request.getIndividuals();
+        Class<?> objClass = getObjClass(individuals);
+        Method idMethod = getMethod(GET_ID, objClass);
+        Map<String, Individual> iMap = getIdToObjMap(individuals
+                .stream().filter(notHavingErrors()).collect(Collectors.toList()), idMethod);
         if (!iMap.isEmpty()) {
             List<String> individualIds = new ArrayList<>(iMap.keySet());
             List<Individual> existingIndividuals = individualRepository.findById(individualIds,
                     getIdFieldName(idMethod), false);
-            List<Individual> individualsWithMismatchedRowVersion =
-                    getEntitiesWithMismatchedRowVersion(iMap, existingIndividuals, idMethod);
-            individualsWithMismatchedRowVersion.forEach(individual -> {
-                Error error = getErrorForRowVersionMismatch();
+            List<Individual> nonExistentIndividuals = checkNonExistentEntities(iMap,
+                    existingIndividuals, idMethod);
+            nonExistentIndividuals.forEach(individual -> {
+                Error error = getErrorForNonExistentEntity();
                 populateErrorDetails(individual, error, errorDetailsMap);
             });
         }
