+++ health-services/individual/src/main/java/org/egov/individual/validators/NullIdValidator.java
@@ -1,8 +1,8 @@
-package org.egov.individual.service;
+package org.egov.individual.validators;
 
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.models.Error;
-import org.egov.common.utils.Validator;
+import org.egov.common.validator.Validator;
 import org.egov.individual.web.models.Individual;
 import org.egov.individual.web.models.IndividualBulkRequest;
 import org.springframework.core.annotation.Order;
@@ -12,6 +12,7 @@ import java.util.List;
 import java.util.Map;
 
 import static org.egov.common.utils.CommonUtils.validateForNullId;
+import static org.egov.individual.Constants.GET_INDIVIDUALS;
 
 @Component
 @Order(value = 1)
@@ -20,6 +21,6 @@ public class NullIdValidator implements Validator<IndividualBulkRequest, Individ
 
     @Override
     public Map<Individual, List<Error>> validate(IndividualBulkRequest request) {
-        return validateForNullId(request, "getIndividuals");
+        return validateForNullId(request, GET_INDIVIDUALS);
     }
 }
