+++ health-services/individual/src/main/java/org/egov/individual/validators/RowVersionValidator.java
@@ -1,8 +1,8 @@
-package org.egov.individual.service;
+package org.egov.individual.validators;
 
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.models.Error;
-import org.egov.common.utils.Validator;
+import org.egov.common.validator.Validator;
 import org.egov.individual.repository.IndividualRepository;
 import org.egov.individual.web.models.Individual;
 import org.egov.individual.web.models.IndividualBulkRequest;
@@ -17,24 +17,23 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static org.egov.common.utils.CommonUtils.checkNonExistentEntities;
+import static org.egov.common.utils.CommonUtils.getEntitiesWithMismatchedRowVersion;
 import static org.egov.common.utils.CommonUtils.getIdFieldName;
+import static org.egov.common.utils.CommonUtils.getIdMethod;
 import static org.egov.common.utils.CommonUtils.getIdToObjMap;
-import static org.egov.common.utils.CommonUtils.getMethod;
-import static org.egov.common.utils.CommonUtils.getObjClass;
 import static org.egov.common.utils.CommonUtils.notHavingErrors;
 import static org.egov.common.utils.CommonUtils.populateErrorDetails;
-import static org.egov.common.utils.ValidatorUtils.getErrorForNonExistentEntity;
+import static org.egov.common.utils.ValidatorUtils.getErrorForRowVersionMismatch;
 
 @Component
-@Order(value = 4)
+@Order(value = 5)
 @Slf4j
-public class NonExistentEntityValidator implements Validator<IndividualBulkRequest, Individual> {
+public class RowVersionValidator implements Validator<IndividualBulkRequest, Individual> {
 
     private final IndividualRepository individualRepository;
 
     @Autowired
-    public NonExistentEntityValidator(IndividualRepository individualRepository) {
+    public RowVersionValidator(IndividualRepository individualRepository) {
         this.individualRepository = individualRepository;
     }
 
@@ -42,19 +41,18 @@ public class NonExistentEntityValidator implements Validator<IndividualBulkReque
     @Override
     public Map<Individual, List<Error>> validate(IndividualBulkRequest request) {
         Map<Individual, List<Error>> errorDetailsMap = new HashMap<>();
-        List<Individual> individuals = request.getIndividuals();
-        Class<?> objClass = getObjClass(individuals);
-        Method idMethod = getMethod("getId", objClass);
-        Map<String, Individual> iMap = getIdToObjMap(individuals
-                .stream().filter(notHavingErrors()).collect(Collectors.toList()), idMethod);
+        Method idMethod = getIdMethod(request.getIndividuals());
+        Map<String, Individual> iMap = getIdToObjMap(request.getIndividuals().stream()
+                .filter(notHavingErrors())
+                .collect(Collectors.toList()), idMethod);
         if (!iMap.isEmpty()) {
             List<String> individualIds = new ArrayList<>(iMap.keySet());
             List<Individual> existingIndividuals = individualRepository.findById(individualIds,
                     getIdFieldName(idMethod), false);
-            List<Individual> nonExistentIndividuals = checkNonExistentEntities(iMap,
-                    existingIndividuals, idMethod);
-            nonExistentIndividuals.forEach(individual -> {
-                Error error = getErrorForNonExistentEntity();
+            List<Individual> individualsWithMismatchedRowVersion =
+                    getEntitiesWithMismatchedRowVersion(iMap, existingIndividuals, idMethod);
+            individualsWithMismatchedRowVersion.forEach(individual -> {
+                Error error = getErrorForRowVersionMismatch();
                 populateErrorDetails(individual, error, errorDetailsMap);
             });
         }
