+++ health-services/individual/src/main/java/org/egov/individual/validators/UniqueEntityValidator.java
@@ -1,8 +1,8 @@
-package org.egov.individual.service;
+package org.egov.individual.validators;
 
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.models.Error;
-import org.egov.common.utils.Validator;
+import org.egov.common.validator.Validator;
 import org.egov.individual.web.models.Individual;
 import org.egov.individual.web.models.IndividualBulkRequest;
 import org.springframework.core.annotation.Order;
