+++ health-services/individual/src/main/resources/application.properties
@@ -53,6 +53,7 @@ kafka.producer.config.buffer_memory_config=33554432
 egov.idgen.host=https://dev.digit.org/
 egov.idgen.path=egov-idgen/id/_generate
 egov.idgen.integration.enabled=true
+idgen.individual.id.format=individual.id
 
 # The value of the following field should be changed to service specific name
 individual.consumer.bulk.create.topic=individual-consumer-bulk-create-topic
@@ -60,6 +61,6 @@ individual.consumer.bulk.update.topic=individual-consumer-bulk-update-topic
 individual.consumer.bulk.delete.topic=individual-consumer-bulk-delete-topic
 individual.producer.save.topic=save-individual-topic
 individual.producer.update.topic=update-individual-topic
-individual.producer.delelte.topic=delete-individual-topic
+individual.producer.delete.topic=delete-individual-topic
