+++ health-services/individual/src/test/java/org/egov/individual/service/IndividualServiceTest.java
@@ -3,11 +3,13 @@ package org.egov.individual.service;
 import org.egov.common.contract.request.RequestInfo;
 import org.egov.common.helper.RequestInfoTestBuilder;
 import org.egov.common.service.IdGenService;
-import org.egov.common.utils.Validator;
+import org.egov.common.validator.Validator;
 import org.egov.individual.config.IndividualProperties;
 import org.egov.individual.helper.IndividualRequestTestBuilder;
 import org.egov.individual.helper.IndividualTestBuilder;
 import org.egov.individual.repository.IndividualRepository;
+import org.egov.individual.validators.AddressTypeValidator;
+import org.egov.individual.validators.UniqueSubEntityValidator;
 import org.egov.individual.web.models.Address;
 import org.egov.individual.web.models.AddressType;
 import org.egov.individual.web.models.Individual;
@@ -29,9 +31,6 @@ import java.util.Collections;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
@@ -64,6 +63,9 @@ class IndividualServiceTest {
     @Mock
     private IndividualProperties properties;
 
+    @Mock
+    private EnrichmentService enrichmentService;
+
     private List<Validator<IndividualBulkRequest, Individual>> validators;
 
     @BeforeEach
@@ -79,170 +81,6 @@ class IndividualServiceTest {
         lenient().when(idGenService.getIdList(any(RequestInfo.class), anyString(),
                 eq(value), eq(null), anyInt()))
                 .thenReturn(Collections.singletonList(o));
-    }
-
-    @Test
-    @DisplayName("should generate address id if address is not null")
-    void shouldGenerateAddressIdIfAddressIsNotNull() throws Exception {
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(IndividualTestBuilder.builder()
-                        .withName()
-                        .withTenantId()
-                        .withAddress()
-                        .build())
-                .build();
-
-        List<Individual> response = individualService.create(request);
-
-        assertNotNull(response.stream().findFirst().get()
-                        .getAddress().stream().findFirst().get()
-                        .getId());
-    }
-
-    @Test
-    @DisplayName("should enrich address audit details and individual id if address is not null")
-    void shouldEnrichAddressWithAuditDetailsAndIndividualIdIfAddressIsNotNull() throws Exception {
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(IndividualTestBuilder.builder()
-                        .withName()
-                        .withTenantId()
-                        .withAddress()
-                        .build())
-                .build();
-
-        List<Individual> response = individualService.create(request);
-
-        assertNotNull(response.stream().findFirst().get()
-                .getAddress().stream().findFirst().get()
-                .getId());
-        assertNotNull(response.stream().findFirst().get()
-                .getAddress().stream().findFirst().get()
-                .getIndividualId());
-        assertNotNull(response.stream().findFirst().get()
-                .getAddress().stream().findFirst().get()
-                .getAuditDetails());
-    }
-
-    @Test
-    @DisplayName("should generate individual id")
-    void shouldGenerateIndividualId() throws Exception {
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(IndividualTestBuilder.builder()
-                        .withTenantId()
-                        .withName()
-                        .build())
-                .build();
-
-        List<Individual> response = individualService.create(request);
-
-        assertEquals("some-individual-id",
-                response.stream().findFirst().get()
-                        .getId());
-    }
-
-    @Test
-    @DisplayName("should enrich individuals")
-    void shouldEnrichIndividuals() throws Exception {
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(IndividualTestBuilder.builder()
-                        .withTenantId()
-                        .withName()
-                        .build())
-                .build();
-
-        List<Individual> response = individualService.create(request);
-
-        assertEquals("some-individual-id",
-                response.stream().findFirst().get()
-                        .getId());
-        assertEquals(1, response.stream().findFirst().get().getRowVersion());
-        assertFalse(response.stream().findFirst().get().getIsDeleted());
-        assertNotNull(response.stream().findFirst().get().getAuditDetails());
-    }
-
-    @Test
-    @DisplayName("should generate identifier if not present")
-    void shouldGenerateIdentifierIfNotPresent() throws Exception {
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(IndividualTestBuilder.builder()
-                        .withTenantId()
-                        .withName()
-                        .build())
-                .build();
-
-        List<Individual> response = individualService.create(request);
-
-        assertNotNull(response.stream().findFirst().get()
-                .getIdentifiers().stream().findFirst().get()
-                .getIdentifierId());
-        assertEquals("SYSTEM_GENERATED",
-                response.stream().findFirst().get()
-                        .getIdentifiers().stream().findFirst().get()
-                        .getIdentifierType());
-    }
-
-    @Test
-    @DisplayName("should enrich identifier with individual id")
-    void shouldEnrichIdentifierWithIndividualId() throws Exception {
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(IndividualTestBuilder.builder()
-                        .withTenantId()
-                        .withName()
-                        .build())
-                .build();
-
-        List<Individual> response = individualService.create(request);
-
-        assertNotNull(response.stream().findFirst().get()
-                .getIdentifiers().stream().findFirst().get()
-                .getIdentifierId());
-        assertEquals("SYSTEM_GENERATED",
-                response.stream().findFirst().get()
-                        .getIdentifiers().stream().findFirst().get()
-                        .getIdentifierType());
-        assertEquals("some-individual-id",
-                response.stream().findFirst().get()
-                        .getIdentifiers().stream().findFirst().get()
-                        .getIndividualId());
-    }
-
-    @Test
-    @DisplayName("should enrich identifiers")
-    void shouldEnrichIdentifiers() throws Exception {
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(IndividualTestBuilder.builder()
-                        .withTenantId()
-                        .withName()
-                        .build())
-                .build();
-
-        List<Individual> response = individualService.create(request);
-
-        assertNotNull(response.stream().findFirst().get()
-                .getIdentifiers().stream().findFirst().get()
-                .getIdentifierId());
-        assertEquals("SYSTEM_GENERATED",
-                response.stream().findFirst().get()
-                        .getIdentifiers().stream().findFirst().get()
-                        .getIdentifierType());
-        assertNotNull(
-                response.stream().findFirst().get()
-                        .getIdentifiers().stream().findFirst().get()
-                        .getId());
-        assertEquals("some-individual-id",
-                response.stream().findFirst().get()
-                        .getIdentifiers().stream().findFirst().get()
-                        .getIdentifierId());
-        assertNotNull(response.stream().findFirst().get()
-                .getIdentifiers().stream().findFirst().get()
-                .getAuditDetails());
     }
 
     @Test
