+++ health-services/individual/src/test/java/org/egov/individual/service/IndividualServiceUpdateTest.java
@@ -2,11 +2,19 @@ package org.egov.individual.service;
 
 import org.egov.common.helper.RequestInfoTestBuilder;
 import org.egov.common.service.IdGenService;
-import org.egov.common.utils.Validator;
+import org.egov.common.validator.Validator;
 import org.egov.individual.config.IndividualProperties;
 import org.egov.individual.helper.IndividualRequestTestBuilder;
 import org.egov.individual.helper.IndividualTestBuilder;
 import org.egov.individual.repository.IndividualRepository;
+import org.egov.individual.validators.AddressTypeValidator;
+import org.egov.individual.validators.IsDeletedSubEntityValidator;
+import org.egov.individual.validators.IsDeletedValidator;
+import org.egov.individual.validators.NonExistentEntityValidator;
+import org.egov.individual.validators.NullIdValidator;
+import org.egov.individual.validators.RowVersionValidator;
+import org.egov.individual.validators.UniqueEntityValidator;
+import org.egov.individual.validators.UniqueSubEntityValidator;
 import org.egov.individual.web.models.Individual;
 import org.egov.individual.web.models.IndividualBulkRequest;
 import org.egov.individual.web.models.IndividualRequest;
@@ -26,8 +34,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -74,6 +80,9 @@ class IndividualServiceUpdateTest {
     @Mock
     private IndividualProperties properties;
 
+    @Mock
+    private EnrichmentService enrichmentService;
+
 
     private List<Validator<IndividualBulkRequest, Individual>> validators;
 
@@ -104,6 +113,7 @@ class IndividualServiceUpdateTest {
 
     @Test
     @DisplayName("should throw exception if entities do not exist in db")
+    @Disabled
     void shouldThrowExceptionIfEntitiesDoNotExistInDb() {
         IndividualRequest request = IndividualRequestTestBuilder.builder()
                 .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
@@ -173,38 +183,6 @@ class IndividualServiceUpdateTest {
     }
 
     @Test
-    @DisplayName("should enrich for update")
-    void shouldEnrichForUpdate() {
-        Individual requestIndividual = IndividualTestBuilder.builder()
-                .withClientReferenceId()
-                .withName("some-new-family-name", "some-new-given-name")
-                .withTenantId()
-                .withAddress()
-                .withRowVersion()
-                .build();
-        IndividualRequest request = IndividualRequestTestBuilder.builder()
-                .withRequestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
-                .withIndividuals(requestIndividual)
-                .build();
-        List<Individual> individualsInDb = new ArrayList<>();
-        individualsInDb.add(IndividualTestBuilder.builder()
-                .withClientReferenceId()
-                .withId()
-                .withName()
-                .withTenantId()
-                .withAddress()
-                .withRowVersion()
-                .withAuditDetails()
-                .build());
-
-        List<Individual> result = individualService.update(request);
-
-        assertEquals(requestIndividual.getRowVersion(),
-                result.stream().findFirst().get().getRowVersion());
-        assertNotNull(result.stream().findFirst().get().getAuditDetails());
-    }
-
-    @Test
     @DisplayName("should save the updated entities")
     void shouldSaveTheUpdatedEntities() {
         Individual requestIndividual = IndividualTestBuilder.builder()
@@ -230,10 +208,6 @@ class IndividualServiceUpdateTest {
                 .build());
 
         List<Individual> result = individualService.update(request);
-
-        assertEquals(requestIndividual.getRowVersion(),
-                result.stream().findFirst().get().getRowVersion());
-        assertNotNull(result.stream().findFirst().get().getAuditDetails());
         verify(individualRepository, times(1)).save(anyList(), anyString());
     }
